tests/wpt/tests/css/css-anchor-position/anchor-size-parse-invalid.html

<!DOCTYPE html>
<title>Tests values that are invalid at parse time for the anchor-size() function</title>
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#anchor-size">
<link rel="author" href="mailto:xiaochengh@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>

<script>
// anchor-size() can only be used in sizing properties
test_invalid_value('margin-top', 'anchor-size(--foo width)');
test_invalid_value('top', 'anchor-size(--foo width)');
test_invalid_value('font-size', 'anchor-size(--foo width)');

// Invalid parameter format
test_invalid_value('width', 'anchor-size(--foo, width)');
test_invalid_value('width', 'anchor-size(--foo width,)');
test_invalid_value('width', 'anchor-size(--foo width height)');
test_invalid_value('width', 'anchor-size(--foo width, 10px 20%)');
test_invalid_value('width', 'anchor-size(--foo width, 10px, 20%)');
test_invalid_value('width', 'anchor-size(--foo,)');
test_invalid_value('width', 'anchor-size(, 10px)');

// Anchor name must be a dashed ident
test_invalid_value('width', 'anchor-size(foo width)');

// Invalid anchor size values
test_invalid_value('width', 'anchor-size(--foo top)');
test_invalid_value('width', 'anchor-size(--foo 10em)');
test_invalid_value('width', 'anchor-size(--foo 100s)');
test_invalid_value('width', 'anchor-size(--foo 50%)');

// Invalid fallback values
test_invalid_value('width', 'anchor-size(--foo width, 1)');
test_invalid_value('width', 'anchor-size(--foo width, 100s)');
test_invalid_value('width', 'anchor-size(--foo width, height)');
test_invalid_value('width', 'anchor-size(--foo width, anchor-size(bar width))');
test_invalid_value('width', 'anchor-size(--foo width, anchor(--bar top))');

// Invalid anchor size values in calc tree
test_invalid_value('width', 'calc(anchor-size(foo width) + 10px + 10%)');
test_invalid_value('width', 'calc(10px + 100 * anchor-size(--foo width, anchor-size(bar bottom)))');
test_invalid_value('width', 'min(anchor-size(--foo width), anchor-size(--bar height), anchor(--baz top))');
</script>
